Integrate Webpack to build frontend assets (#29)
* add intial setup for webpack pipeline
* add external depencies

* remove test component

* fix issues with eslint jshint setup

* replace legacy pipeline references in view with webpack

* add package-lock.json (from latest npm version)

* ensure we server from the same port as defined in development

* add modernizr via webpack

* add webpack based modernizr

* require CC classes as needed

…as they are no longer available on the global scope in webpack

* require dependencies where needed is classcentraljs, ensure constants are defined

* use handlebars from npm, use alias to address webpack issue with inclusion

* add import, expose, and export loaders for certain files

* define this as window to avoid webpack issues

* include necessary files in main Entry

* use https

* use https

* include and load dependencies via npm where we can

* add package-log

* remove files now included via npm

* explicity set dependencies

* remove stray debugger

* fix string to integer conversion with elasticsearch script

* update client and build scripts to use specific configs

* separate out dev and prod webpack configs

* set up config yml files to webpack asset inclusion

* use webpack assets in base twig template

* remove duplicate config options

* add build files

* fix asset urls

* udpate build assets

* update asset hash

* commit bug fixes

* update to latest hash

* fix missing cc class definitions

* fix loading issue with follow script

* fix scope issues with cc classes and flexslider

* fix utilities references

* fix profile references

* Remove hardcoding of bundled css and js

Read the bundled css and json file names from the manifest

* fix countup scope issue

* fix asset loading bug

* add webpack clean plugin on production builds to remove unsued assets

* fix pnotify references

* latest build

* update packagejson versions

* Tweaked notification show time

* Server prod assets by default

To use the webpack dev server, now you need to set the webpack_dev_server in parameters.yml to true

+++ src/ClassCentral/ElasticSearchBundle/Finder.php
@@ -165,7 +165,7 @@ class Finder {
             )
         );
 
-        
+
 
         if( !empty($mustNot) )
         {
@@ -231,7 +231,8 @@ class Finder {
                         // Number of sessions
 
                         numSessionsScore = numSessions*20;
-                        return _score*(ratingScore + followedScore + startingSoonScore + newScore - numSessionsScore  +  1);
+
+                        return _score*((int)ratingScore + (int)followedScore + (int)startingSoonScore + (int)newScore - (int)numSessionsScore  +  1);
                     "
                  )
         ));
@@ -366,4 +367,4 @@ class Finder {
 
         return $facets;
     }
-} 
+}
